backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
django-ninja>=1.1.0,<2.0.0
django-cors-headers>=4.0.0   
djangorestframework
python-dotenv
lucide-react
Pillow